pyproject.toml
[project]
name = "histograph"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.0",
    "python-dotenv>=1.0",
    "openai>=1.0",
]

[project.scripts]
fetch-videos = "fetch_videos:main"
parse-videos = "parse_videos:main"
migrate-db   = "migrate:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
